PR: 34241 Tar task now accepts files <8GB (was <2GB), according to POSIX tar standard.
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@278407 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/tar/TarOutputStream.java
@@ -43,8 +43,8 @@ public class TarOutputStream extends FilterOutputStream {
     public static final int LONGFILE_GNU = 2;
 
     protected boolean   debug;
-    protected int       currSize;
-    protected int       currBytes;
+    protected long      currSize;
+    protected long      currBytes;
     protected byte[]    oneBuf;
     protected byte[]    recordBuf;
     protected int       assemLen;
@@ -189,7 +189,7 @@ public class TarOutputStream extends FilterOutputStream {
         if (entry.isDirectory()) {
             this.currSize = 0;
         } else {
-            this.currSize = (int) entry.getSize();
+            this.currSize = entry.getSize();
         }
     }
